pythonanywhere_requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
email-validator==2.0.0
numpy==1.25.2
Pillow==10.1.0
Werkzeug==2.3.7
gunicorn==23.0.0
opencv-python==4.8.1.78
pandas==2.1.1
face-recognition==1.3.0
dlib==19.24.0
pytz==2023.3
psycopg2-binary==2.9.9
torch==2.1.0
torchvision==0.16.0